REVERTING TO LAST GOOD STATE
Downloaded previously commited files and copied over to the empty cms
directory. Now going to commit em, and push. Let's see what's going to
happen. PS: If anyone is going to see this in the future, please leave a
comment on how would you fuckin restore a previous state of all the
files in the repository. No fancy stuff. I want back everything as it
was in a prev commit!